# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "vms Api Documentation",
    "description": "vms Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "vms Logo",
      "altText": "vms Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "vms API Support",
      "url": "https://vms.com/info/support",
      "email": "support@vms.com"
    },
    "termsOfService": "https://vms.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/services": {
      "get": {
        "summary": "List Services",
        "description": "Services",
        "operationId": "ListServices",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListServices"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/services/view/{id}": {
      "get": {
        "summary": "View Services",
        "description": "Vue Service",
        "operationId": "ViewServices",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewServices"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/services/add": {
      "post": {
        "summary": "Add Services",
        "description": "Ajouter un nouveau",
        "operationId": "AddServices",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Services"
        ],
        
        "requestBody": {
          "description": "Add Services",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddServices"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewServices"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/services/edit/{id}": {
      "put": {
        "summary": "Edit Services",
        "description": "modifier",
        "operationId": "EditServices",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Services",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditServices"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditServices"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/services/delete/{id}": {
      "delete": {
        "summary": "Delete Services",
        "description": "Services",
        "operationId": "DeleteServices",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListServices": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          }
        }
      },
      "ViewServices": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          }
        }
      },
      "AddServices": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          }
        }
      },
      "EditServices": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "type": {
            "description": "Type",
            "type": "string"
          },
          "montant": {
            "description": "Montant",
            "type": "string"
          },
          "fournisseur": {
            "description": "Fournisseur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "image": {
            "description": "Image",
            "type": "string"
          },
          "document": {
            "description": "Document",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}